Reshuffle colors, improve accuracy of cspace conversion
Former-commit-id: 3c378bc93b0edd73d1c8f07fab33b1943b9d4411
Former-commit-id: 2fa968ca7c3535ab92251d55afba20e64c574089
Former-commit-id: 12065bfba7ce14e175d99d15d2f0a937665fa395

+++ src/ImageProcessor/Samplers/Resamplers/TriangleResampler.cs
@@ -11,10 +11,10 @@ namespace ImageProcessor.Samplers
     public class TriangleResampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 1;
+        public float Radius => 1;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
             if (x < 0)
             {
